Crate libffi_sys

Low-level Rust bindings for libffi

The C libffi library provides two main facilities: assembling calls to functions dynamically, and creating closures that can be called as ordinary C functions. This is an undocumented wrapper, generated by bindgen, intended as the basis for higher-level bindings, but you can see the C libffi documentation.

See the libffi crate for a higher-level API.

Usage

It’s on crates.io, but before you build it, make sure you have the dependencies installed first:

An up-to-date version of C libffi Version 3.2.1 is known to work. Earlier versions, such as the versions that come with Mac OS and Fedora, are known not to; neither will the version installed by Homebrew (3.0.13).

pkg-config, which you probably already have if you’re on Linux. For Mac users, the version installed by Homebrew is up to date. (I don’t know how this works on Windows; contact me if you’d like to help figure it out.)

Then add

[dependencies]
libffi-sys = "0.2"

to your Cargo.toml and

extern crate libffi;

to your crate root.

Structs

_ffi_type

ffi_cif

ffi_closure

ffi_java_raw_closure

ffi_raw

ffi_raw_closure

Enums

ffi_abi

ffi_status

ffi_type_enum

Constants

FFI_DEFAULT_ABI

Statics

ffi_type_complex_double

ffi_type_complex_float

ffi_type_complex_longdouble

ffi_type_double

ffi_type_float

ffi_type_longdouble

ffi_type_pointer

ffi_type_sint16

ffi_type_sint32

ffi_type_sint64

ffi_type_sint8

ffi_type_uint16

ffi_type_uint32

ffi_type_uint64

ffi_type_uint8

ffi_type_void

Functions

ffi_call

ffi_closure_alloc

ffi_closure_free

ffi_java_ptrarray_to_raw

ffi_java_raw_call

ffi_java_raw_size

ffi_java_raw_to_ptrarray

ffi_prep_cif

ffi_prep_cif_core

ffi_prep_cif_var

ffi_prep_closure

ffi_prep_closure_loc

ffi_prep_java_raw_closure

ffi_prep_java_raw_closure_loc

ffi_prep_raw_closure

ffi_prep_raw_closure_loc

ffi_ptrarray_to_raw

ffi_raw_call

ffi_raw_size

ffi_raw_to_ptrarray

Type Definitions

FFI_TYPE

ffi_arg

ffi_java_raw

ffi_sarg

ffi_type

max_align_t

ptrdiff_t

size_t

wchar_t