Crate libffi_sys

Low-level Rust bindings for libffi

The C libffi library provides two main facilities: assembling calls to functions dynamically, and creating closures that can be called as ordinary C functions. This is an undocumented wrapper, generated by bindgen, intended as the basis for higher-level bindings, but you can see the C libffi documentation.

See the libffi crate for a higher-level API.

Usage

Make sure you have a working make, C compiler, automake, and autoconf first. It’s on crates.io, so you can add

[dependencies]
libffi-sys = "0.5.2"

to your Cargo.toml and

extern crate libffi_sys;

to your crate root.

Structs

__BindgenUnionField

_ffi_type

ffi_cif

ffi_closure

ffi_java_raw_closure

ffi_raw

ffi_raw_closure

Enums

ffi_abi

ffi_status

ffi_type_enum

Constants

ARG_MAX

BC_BASE_MAX

BC_DIM_MAX

BC_SCALE_MAX

BC_STRING_MAX

CHARCLASS_NAME_MAX

CHAR_BIT

CHAR_MAX

CHAR_MIN

CHILD_MAX

CLK_TCK

COLL_WEIGHTS_MAX

EQUIV_CLASS_MAX

EXPR_NEST_MAX

FFI_64_BIT_MAX

FFI_CLOSURES

FFI_NATIVE_RAW_API

FFI_SIZEOF_ARG

FFI_SIZEOF_JAVA_RAW

FFI_TRAMPOLINE_SIZE

FFI_TYPE_COMPLEX

FFI_TYPE_DOUBLE

FFI_TYPE_FLOAT

FFI_TYPE_INT

FFI_TYPE_LAST

FFI_TYPE_LONGDOUBLE

FFI_TYPE_POINTER

FFI_TYPE_SINT8

FFI_TYPE_SINT16

FFI_TYPE_SINT32

FFI_TYPE_SINT64

FFI_TYPE_STRUCT

FFI_TYPE_UINT8

FFI_TYPE_UINT16

FFI_TYPE_UINT32

FFI_TYPE_UINT64

FFI_TYPE_VOID

GID_MAX

INT_MAX

INT_MIN

IOV_MAX

LINE_MAX

LINK_MAX

LLONG_MAX

LLONG_MIN

LONG_BIT

LONG_MAX

LONG_MIN

MAX_CANON

MAX_INPUT

MB_LEN_MAX

NAME_MAX

NGROUPS_MAX

NL_ARGMAX

NL_LANGMAX

NL_MSGMAX

NL_NMAX

NL_SETMAX

NL_TEXTMAX

NZERO

OFF_MAX

OFF_MIN

OPEN_MAX

PASS_MAX

PATH_MAX

PIPE_BUF

PTHREAD_DESTRUCTOR_ITERATIONS

PTHREAD_KEYS_MAX

PTHREAD_STACK_MIN

QUAD_MAX

QUAD_MIN

RE_DUP_MAX

SCHAR_MAX

SCHAR_MIN

SHRT_MAX

SHRT_MIN

SIZE_T_MAX

SSIZE_MAX

UCHAR_MAX

UID_MAX

UINT_MAX

ULLONG_MAX

ULONG_MAX

UQUAD_MAX

USHRT_MAX

WORD_BIT

_DARWIN_FEATURE_64_BIT_INODE

_DARWIN_FEATURE_ONLY_UNIX_CONFORMANCE

_DARWIN_FEATURE_UNIX_CONFORMANCE

_POSIX2_BC_BASE_MAX

_POSIX2_BC_DIM_MAX

_POSIX2_BC_SCALE_MAX

_POSIX2_BC_STRING_MAX

_POSIX2_CHARCLASS_NAME_MAX

_POSIX2_COLL_WEIGHTS_MAX

_POSIX2_EQUIV_CLASS_MAX

_POSIX2_EXPR_NEST_MAX

_POSIX2_LINE_MAX

_POSIX2_RE_DUP_MAX

_POSIX_AIO_LISTIO_MAX

_POSIX_AIO_MAX

_POSIX_ARG_MAX

_POSIX_CHILD_MAX

_POSIX_CLOCKRES_MIN

_POSIX_DELAYTIMER_MAX

_POSIX_HOST_NAME_MAX

_POSIX_LINK_MAX

_POSIX_LOGIN_NAME_MAX

_POSIX_MAX_CANON

_POSIX_MAX_INPUT

_POSIX_MQ_OPEN_MAX

_POSIX_MQ_PRIO_MAX

_POSIX_NAME_MAX

_POSIX_NGROUPS_MAX

_POSIX_OPEN_MAX

_POSIX_PATH_MAX

_POSIX_PIPE_BUF

_POSIX_RE_DUP_MAX

_POSIX_RTSIG_MAX

_POSIX_SEM_NSEMS_MAX

_POSIX_SEM_VALUE_MAX

_POSIX_SIGQUEUE_MAX

_POSIX_SSIZE_MAX

_POSIX_SS_REPL_MAX

_POSIX_STREAM_MAX

_POSIX_SYMLINK_MAX

_POSIX_SYMLOOP_MAX

_POSIX_THREAD_DESTRUCTOR_ITERATIONS

_POSIX_THREAD_KEYS_MAX

_POSIX_THREAD_THREADS_MAX

_POSIX_TIMER_MAX

_POSIX_TRACE_EVENT_NAME_MAX

_POSIX_TRACE_NAME_MAX

_POSIX_TRACE_SYS_MAX

_POSIX_TRACE_USER_EVENT_MAX

_POSIX_TTY_NAME_MAX

_POSIX_TZNAME_MAX

_XOPEN_IOV_MAX

_XOPEN_NAME_MAX

_XOPEN_PATH_MAX

__DARWIN_64_BIT_INO_T

__DARWIN_CLK_TCK

__DARWIN_C_ANSI

__DARWIN_C_FULL

__DARWIN_C_LEVEL

__DARWIN_NON_CANCELABLE

__DARWIN_ONLY_64_BIT_INO_T

__DARWIN_ONLY_UNIX_CONFORMANCE

__DARWIN_ONLY_VERS_1050

__DARWIN_SUF_1050

__DARWIN_SUF_64_BIT_INO_T

__DARWIN_SUF_EXTSN

__DARWIN_UNIX03

__DARWIN_VERS_1050

__STDC_WANT_LIB_EXT1__

ffi_abi_FFI_DEFAULT_ABI

Statics

ffi_type_complex_double

ffi_type_complex_float

ffi_type_complex_longdouble

ffi_type_double

ffi_type_float

ffi_type_longdouble

ffi_type_pointer

ffi_type_sint8

ffi_type_sint16

ffi_type_sint32

ffi_type_sint64

ffi_type_uint8

ffi_type_uint16

ffi_type_uint32

ffi_type_uint64

ffi_type_void

Functions

ffi_call⚠

ffi_closure_alloc⚠

ffi_closure_free⚠

ffi_java_ptrarray_to_raw⚠

ffi_java_raw_call⚠

ffi_java_raw_size⚠

ffi_java_raw_to_ptrarray⚠

ffi_prep_cif⚠

ffi_prep_cif_core⚠

ffi_prep_cif_var⚠

ffi_prep_closure⚠

ffi_prep_closure_loc⚠

ffi_prep_java_raw_closure⚠

ffi_prep_java_raw_closure_loc⚠

ffi_prep_raw_closure⚠

ffi_prep_raw_closure_loc⚠

ffi_ptrarray_to_raw⚠

ffi_raw_call⚠

ffi_raw_size⚠

ffi_raw_to_ptrarray⚠

Type Definitions

FFI_TYPE

ffi_arg

ffi_java_raw

ffi_sarg

ffi_type

wchar_t